requirements.txt
flask==3.0.0
flask-sqlalchemy==3.1.1
flask-jwt-extended==4.6.0
flask-cors==4.0.0
flask-migrate==4.0.5
flask-restx==1.3.0
python-dotenv==1.0.0
psycopg2-binary==2.9.9
psycopg2==2.9.9
redis==5.0.1
celery==5.3.4
gunicorn==21.2.0
pytest==7.4.3
pytest-flask==1.3.0
black==23.11.0
flake8==6.1.0
python-multipart==0.0.6
bcrypt==4.1.2
email-validator==2.1.0
marshmallow==3.20.1
flask-limiter==3.5.0
requests==2.31.0
schedule==1.2.0